# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PasswordEntryStrength = typeof PasswordEntryStrength[keyof typeof PasswordEntryStrength];


export const PasswordEntryStrength = {
  weak: 'weak',
  medium: 'medium',
  strong: 'strong',
} as const;

export interface PasswordEntry {
  id: number;
  title: string;
  username: string;
  password: string;
  /** @nullable */
  url?: string | null;
  category: string;
  /** @nullable */
  notes?: string | null;
  strength?: PasswordEntryStrength;
  createdAt: string;
  updatedAt: string;
}

export interface PasswordInput {
  /** @minLength 1 */
  title: string;
  username: string;
  /** @minLength 1 */
  password: string;
  url?: string;
  category: string;
  notes?: string;
}

export interface PasswordUpdate {
  /** @minLength 1 */
  title?: string;
  username?: string;
  password?: string;
  url?: string;
  category?: string;
  notes?: string;
}

export interface PasswordStats {
  total: number;
  weak: number;
  medium: number;
  strong: number;
  reused: number;
}

export interface Document {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  notes?: string | null;
  isExpired?: boolean;
  /** @nullable */
  daysUntilExpiry?: number | null;
  createdAt: string;
}

export interface DocumentInput {
  /** @minLength 1 */
  name: string;
  category: string;
  fileUrl?: string;
  expiryDate?: string;
  notes?: string;
}

export interface DocumentUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  fileUrl?: string;
  expiryDate?: string;
  notes?: string;
}

export type FinanceRecordType = typeof FinanceRecordType[keyof typeof FinanceRecordType];


export const FinanceRecordType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface FinanceRecord {
  id: number;
  type: FinanceRecordType;
  category: string;
  amount: number;
  /** @nullable */
  description?: string | null;
  date: string;
  createdAt: string;
}

export type FinanceInputType = typeof FinanceInputType[keyof typeof FinanceInputType];


export const FinanceInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface FinanceInput {
  type: FinanceInputType;
  category: string;
  /** @minimum 0.01 */
  amount: number;
  description?: string;
  date: string;
}

export type FinanceUpdateType = typeof FinanceUpdateType[keyof typeof FinanceUpdateType];


export const FinanceUpdateType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface FinanceUpdate {
  type?: FinanceUpdateType;
  category?: string;
  amount?: number;
  description?: string;
  date?: string;
}

export interface CategoryBreakdown {
  category: string;
  amount: number;
  percentage: number;
}

export interface FinanceSummary {
  month: number;
  year: number;
  totalIncome: number;
  totalExpenses: number;
  savings: number;
  savingsRate: number;
  byCategory: CategoryBreakdown[];
}

export interface DashboardOverview {
  totalPasswords: number;
  totalDocuments: number;
  expiringDocuments: number;
  monthlyBalance: number;
  weakPasswords: number;
  recentAlerts: number;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  weak_password: 'weak_password',
  expiring_document: 'expiring_document',
  overspending: 'overspending',
  expired_document: 'expired_document',
} as const;

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Alert {
  id: number;
  type: AlertType;
  severity: AlertSeverity;
  title: string;
  message: string;
  createdAt: string;
}

export type ListPasswordsParams = {
search?: string;
category?: string;
};

export type ListDocumentsParams = {
category?: string;
};

export type ListFinanceRecordsParams = {
type?: ListFinanceRecordsType;
month?: string;
year?: number;
};

export type ListFinanceRecordsType = typeof ListFinanceRecordsType[keyof typeof ListFinanceRecordsType];


export const ListFinanceRecordsType = {
  income: 'income',
  expense: 'expense',
} as const;

export type GetFinanceSummaryParams = {
month?: number;
year?: number;
};