# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Key Wallet API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: passwords
    description: Password vault operations
  - name: documents
    description: Document vault operations
  - name: finance
    description: Finance tracker operations
  - name: dashboard
    description: Dashboard and insights
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /passwords:
    get:
      operationId: listPasswords
      tags: [passwords]
      summary: List all password entries
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: category
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of passwords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PasswordEntry"
    post:
      operationId: createPassword
      tags: [passwords]
      summary: Create a new password entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordEntry"

  /passwords/{id}:
    get:
      operationId: getPassword
      tags: [passwords]
      summary: Get a password entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Password entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordEntry"
        "404":
          description: Not found
    patch:
      operationId: updatePassword
      tags: [passwords]
      summary: Update a password entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordEntry"
    delete:
      operationId: deletePassword
      tags: [passwords]
      summary: Delete a password entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /passwords/stats:
    get:
      operationId: getPasswordStats
      tags: [passwords]
      summary: Get password health statistics
      responses:
        "200":
          description: Password statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordStats"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /finance/records:
    get:
      operationId: listFinanceRecords
      tags: [finance]
      summary: List finance records
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense]
          required: false
        - in: query
          name: month
          schema:
            type: string
          required: false
        - in: query
          name: year
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of finance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceRecord"
    post:
      operationId: createFinanceRecord
      tags: [finance]
      summary: Create a finance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceRecord"

  /finance/records/{id}:
    patch:
      operationId: updateFinanceRecord
      tags: [finance]
      summary: Update a finance record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceRecord"
    delete:
      operationId: deleteFinanceRecord
      tags: [finance]
      summary: Delete a finance record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /finance/summary:
    get:
      operationId: getFinanceSummary
      tags: [finance]
      summary: Get monthly finance summary
      parameters:
        - in: query
          name: month
          schema:
            type: integer
          required: false
        - in: query
          name: year
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Finance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceSummary"

  /dashboard/overview:
    get:
      operationId: getDashboardOverview
      tags: [dashboard]
      summary: Get dashboard overview stats
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverview"

  /dashboard/alerts:
    get:
      operationId: getDashboardAlerts
      tags: [dashboard]
      summary: Get active security and finance alerts
      responses:
        "200":
          description: Active alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    PasswordEntry:
      type: object
      required: [id, title, username, password, category, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        username:
          type: string
        password:
          type: string
        url:
          type: ["string", "null"]
        category:
          type: string
        notes:
          type: ["string", "null"]
        strength:
          type: string
          enum: [weak, medium, strong]
        createdAt:
          type: string
        updatedAt:
          type: string

    PasswordInput:
      type: object
      required: [title, username, password, category]
      properties:
        title:
          type: string
          minLength: 1
        username:
          type: string
        password:
          type: string
          minLength: 1
        url:
          type: string
        category:
          type: string
        notes:
          type: string

    PasswordUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        username:
          type: string
        password:
          type: string
        url:
          type: string
        category:
          type: string
        notes:
          type: string

    PasswordStats:
      type: object
      required: [total, weak, medium, strong, reused]
      properties:
        total:
          type: integer
        weak:
          type: integer
        medium:
          type: integer
        strong:
          type: integer
        reused:
          type: integer

    Document:
      type: object
      required: [id, name, category, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        fileUrl:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isExpired:
          type: boolean
        daysUntilExpiry:
          type: ["integer", "null"]
        createdAt:
          type: string

    DocumentInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        fileUrl:
          type: string
        expiryDate:
          type: string
        notes:
          type: string

    DocumentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        fileUrl:
          type: string
        expiryDate:
          type: string
        notes:
          type: string

    FinanceRecord:
      type: object
      required: [id, type, category, amount, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        amount:
          type: number
        description:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string

    FinanceInput:
      type: object
      required: [type, category, amount, date]
      properties:
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
        date:
          type: string

    FinanceUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        date:
          type: string

    FinanceSummary:
      type: object
      required: [month, year, totalIncome, totalExpenses, savings, savingsRate, byCategory]
      properties:
        month:
          type: integer
        year:
          type: integer
        totalIncome:
          type: number
        totalExpenses:
          type: number
        savings:
          type: number
        savingsRate:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"

    CategoryBreakdown:
      type: object
      required: [category, amount, percentage]
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number

    DashboardOverview:
      type: object
      required: [totalPasswords, totalDocuments, expiringDocuments, monthlyBalance, weakPasswords, recentAlerts]
      properties:
        totalPasswords:
          type: integer
        totalDocuments:
          type: integer
        expiringDocuments:
          type: integer
        monthlyBalance:
          type: number
        weakPasswords:
          type: integer
        recentAlerts:
          type: integer

    Alert:
      type: object
      required: [id, type, severity, title, message, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [weak_password, expiring_document, overspending, expired_document]
        severity:
          type: string
          enum: [low, medium, high]
        title:
          type: string
        message:
          type: string
        createdAt:
          type: string