# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  notes?: string | null;
  isExpired?: boolean;
  /** @nullable */
  daysUntilExpiry?: number | null;
  createdAt: string;
}