# lib/api-zod/src/generated/types/financeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { FinanceInputType } from './financeInputType';

export interface FinanceInput {
  type: FinanceInputType;
  category: string;
  /** @minimum 0.01 */
  amount: number;
  description?: string;
  date: string;
}