# lib/api-zod/src/generated/types/financeRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { FinanceRecordType } from './financeRecordType';

export interface FinanceRecord {
  id: number;
  type: FinanceRecordType;
  category: string;
  amount: number;
  /** @nullable */
  description?: string | null;
  date: string;
  createdAt: string;
}