# lib/api-zod/src/generated/types/financeSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryBreakdown } from './categoryBreakdown';

export interface FinanceSummary {
  month: number;
  year: number;
  totalIncome: number;
  totalExpenses: number;
  savings: number;
  savingsRate: number;
  byCategory: CategoryBreakdown[];
}