# lib/api-zod/src/generated/types/passwordEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordEntryStrength } from './passwordEntryStrength';

export interface PasswordEntry {
  id: number;
  title: string;
  username: string;
  password: string;
  /** @nullable */
  url?: string | null;
  category: string;
  /** @nullable */
  notes?: string | null;
  strength?: PasswordEntryStrength;
  createdAt: string;
  updatedAt: string;
}