# lib/api-zod/src/generated/types/passwordStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordStats {
  total: number;
  weak: number;
  medium: number;
  strong: number;
  reused: number;
}