# lib/api-zod/src/generated/types/passwordUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Key Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordUpdate {
  /** @minLength 1 */
  title?: string;
  username?: string;
  password?: string;
  url?: string;
  category?: string;
  notes?: string;
}